fix(*): move utils
